* 读csv会忽略BOM数据 [Issue #3137](https://github.com/alibaba/easyexcel/issues/3137) * 解决csv用office打开乱码的问题，写csv默认带上BOM数据

+++ easyexcel-test/src/test/java/com/alibaba/easyexcel/test/core/bom/BomData.java
@@ -1,6 +1,7 @@
 package com.alibaba.easyexcel.test.core.bom;
 
 import com.alibaba.excel.annotation.ExcelProperty;
+
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,5 +13,5 @@ public class BomData {
     @ExcelProperty("姓名")
     private String name;
     @ExcelProperty("年纪")
-    private Integer age;
+    private Long age;
 }
